--- a/ui/.eslintrc.base.js
+++ b/ui/.eslintrc.base.js
@@ -67,4 +67,6 @@ module.exports = {
     // Not necessary in React 17
     'react/react-in-jsx-scope': 'off',
   },
+
+  ignorePatterns: ['**/dist'],
 };