--- a/ui/.gitignore
+++ b/ui/.gitignore
@@ -24,8 +24,9 @@ yarn-error.log*
 # Build output
 dist
 
-# ignore file copied during the release process of the UI
+# Ignore turborepo cache
+.turbo
 
+# ignore file copied during the release process of the UI
 LICENSE
 CHANGELOG.md
-