--- a/ui/app/webpack.prod.ts
+++ b/ui/app/webpack.prod.ts
@@ -13,12 +13,16 @@
 
 import { Configuration } from 'webpack';
 import { merge } from 'webpack-merge';
+import { ESBuildMinifyPlugin } from 'esbuild-loader';
 import { commonConfig } from './webpack.common';
 
 const prodConfig: Configuration = {
   mode: 'production',
   bail: true,
   devtool: 'source-map',
+  optimization: {
+    minimizer: [new ESBuildMinifyPlugin({ target: 'es2018' })],
+  },
 };
 
 const merged = merge(commonConfig, prodConfig);