--- a/ui/components/package.json
+++ b/ui/components/package.json
@@ -12,13 +12,12 @@
     "url": "https://github.com/perses/perses/issues"
   },
   "module": "dist/index.js",
-  "main": "dist/cjs/index.js",
+  "main": "dist/index.cjs.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "clean": "rimraf dist/",
-    "build": "tsc",
-    "build:cjs": "tsc --project ./tsconfig.cjs.json",
-    "start": "tsc --watch --preserveWatchOutput",
+    "build": "tsup --config ../tsup.config.ts",
+    "start": "npm run build -- --watch",
     "test": "TZ=UTC jest",
     "test:watch": "TZ=UTC jest --watch",
     "lint": "eslint src --ext .ts,.tsx",