--- a/ui/components/tsconfig.cjs.json
+++ b/ui/components/tsconfig.cjs.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.cjs.json",
-  "include": ["src"],
-  "compilerOptions": {
-    "outDir": "./dist/cjs",
-    "rootDir": "./src"
-  }
-}