--- a/ui/components/tsconfig.json
+++ b/ui/components/tsconfig.json
@@ -3,13 +3,6 @@
   "include": ["src"],
   "compilerOptions": {
     "outDir": "./dist",
-    "rootDir": "./src",
-    "tsBuildInfoFile": "../.build-cache/components.tsbuildinfo"
-  },
-  "references": [
-    {
-      "path": "../core",
-      "prepend": false
-    }
-  ]
+    "rootDir": "./src"
+  }
 }