--- a/ui/dashboards/package.json
+++ b/ui/dashboards/package.json
@@ -16,8 +16,9 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "clean": "rimraf dist/",
-    "build": "tsc --build",
+    "build": "tsc",
     "build:cjs": "tsc --project ./tsconfig.cjs.json",
+    "start": "tsc --watch",
     "test": "TZ=UTC jest",
     "test:watch": "TZ=UTC jest --watch",
     "lint": "eslint src --ext .ts,.tsx",