--- a/ui/dashboards/package.json
+++ b/ui/dashboards/package.json
@@ -12,15 +12,14 @@
     "url": "https://github.com/perses/perses/issues"
   },
   "module": "dist/index.js",
-  "main": "dist/cjs/index.js",
+  "main": "dist/index.cjs.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "clean": "rimraf dist/",
-    "build": "tsc",
-    "build:cjs": "tsc --project ./tsconfig.cjs.json",
-    "start": "tsc --watch --preserveWatchOutput",
+    "build": "tsup --config ../tsup.config.ts",
+    "start": "npm run build -- --watch",
     "test": "TZ=UTC jest",
-    "test:watch": "TZ=UTC jest --watch",
+    "test:watch": "npm run test -- --watch",
     "lint": "eslint src --ext .ts,.tsx",
     "lint:fix": "eslint --fix src --ext .ts,.tsx"
   },