--- a/ui/dashboards/package.json
+++ b/ui/dashboards/package.json
@@ -29,6 +29,7 @@
     "@perses-dev/plugin-system": "^0.7.1",
     "@types/react-grid-layout": "^1.3.2",
     "immer": "^9.0.15",
+    "mdi-material-ui": "^7.4.0",
     "react-grid-layout": "^1.3.4",
     "react-intersection-observer": "^9.4.0",
     "use-immer": "^0.7.0",