--- a/ui/dashboards/package.json
+++ b/ui/dashboards/package.json
@@ -28,6 +28,7 @@
     "@perses-dev/core": "^0.7.1",
     "@perses-dev/plugin-system": "^0.7.1",
     "@types/react-grid-layout": "^1.3.2",
+    "date-fns": "^2.28.0",
     "immer": "^9.0.15",
     "mdi-material-ui": "^7.4.0",
     "react-grid-layout": "^1.3.4",