--- a/ui/dashboards/tsconfig.cjs.json
+++ b/ui/dashboards/tsconfig.cjs.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.cjs.json",
-  "include": ["src"],
-  "compilerOptions": {
-    "outDir": "./dist/cjs",
-    "rootDir": "./src"
-  }
-}