--- a/ui/package.json
+++ b/ui/package.json
@@ -11,7 +11,8 @@
     "lint:fix": "npm run lint:fix --workspaces",
     "start": "turbo run start",
     "clean": "turbo run clean",
-    "reinstall": "npm run clean && rm -rf **/node_modules && npm install"
+    "reinstall": "npm run clean && rm -rf **/node_modules && npm install",
+    "clear-cache": "rimraf node_modules/.cache/turbo"
   },
   "workspaces": [
     "app",
@@ -46,6 +47,7 @@
     "react-dom": "^18.0.0",
     "rimraf": "^3.0.2",
     "ts-jest": "^27.1.5",
+    "tsup": "^6.2.3",
     "turbo": "^1.5.2",
     "typescript": "^4.7.4"
   }