--- a/ui/turbo.json
+++ b/ui/turbo.json
@@ -1,0 +1,24 @@
+{
+  "$schema": "https://turborepo.org/schema.json",
+  "pipeline": {
+    // The build task depends on all its upstream dependencies build tasks being run first
+    "build": {
+      "dependsOn": ["^build"],
+      "outputs": ["dist/**"]
+    },
+    "clean": {
+      "dependsOn": []
+    },
+    "lint": {
+      "dependsOn": [],
+      "outputs": []
+    },
+    "start": {
+      "dependsOn": []
+    },
+    "test": {
+      "dependsOn": [],
+      "outputs": []
+    }
+  }
+}