--- a/ui/turbo.json
+++ b/ui/turbo.json
@@ -7,14 +7,16 @@
       "outputs": ["dist/**"]
     },
     "clean": {
-      "dependsOn": []
+      "dependsOn": [],
+      "cache": false
     },
     "lint": {
       "dependsOn": [],
       "outputs": []
     },
     "start": {
-      "dependsOn": [],
+      // Make sure all packages are built once before starting watch mode
+      "dependsOn": ["^build"],
       // Don't cache the output of the dev/start task
       "cache": false
     },