… ui/components/tsconfig.tsbuildinfo
… ui/core/tsconfig.tsbuildinfo
… ui/dashboards/tsconfig.tsbuildinfo
… ui/panels-plugin/tsconfig.tsbuildinfo
… ui/plugin-system/tsconfig.tsbuildinfo
… ui/prometheus-plugin/tsconfig.tsbuildinfo
… ui/tsconfig.cjs.json