@LukeTillman 2022-09-24
The CI results look promising. Here's a [recent run before the change](https://github.com/perses/perses/actions/runs/3109414385/jobs/5039730011) where it looks like the 2 UI build tasks took about 2 minutes total to finish. It looks like it takes about 1 minute now in this PR.